fix task solution #2
