+++ src/components/TabsPage.tsx
@@ -1,36 +1,10 @@
-import { Link, useParams } from 'react-router-dom';
-
-const tabs = [
-  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-];
+import { Tabs } from './Tabs';
 
 export const TabsPage = () => {
-  const { tabId } = useParams();
-
-  const activeTab = tabs.find(t => t.id === tabId);
-
   return (
     <>
       <h1 className="title">Tabs page</h1>
-      <div className="tabs is-boxed">
-        <ul>
-          {tabs.map(tab => (
-            <li
-              data-cy="Tab"
-              className={tabId === tab.id ? 'is-active' : ''}
-              key={tab.id}
-            >
-              <Link to={`/tabs/${tab.id}`}>{tab.title}</Link>
-            </li>
-          ))}
-        </ul>
-      </div>
-
-      <div className="block" data-cy="TabContent">
-        {activeTab ? activeTab.content : 'Please select a tab'}
-      </div>
+      <Tabs />
     </>
   );
 };
